+++ test/utils-test.js
@@ -27,6 +27,16 @@ describe('Utils', function () {
       bar: 1,
       foo: 1
     }
+    var test_query3 = {
+      foo: 1,
+      bar: 2,
+      where: 'where clause1'
+    }
+    var test_query2 = {
+      foo: 1,
+      bar: 2,
+      where: 'where clause2'
+    }
     var expected_key = Utils.createCacheKey(params, reference_query)
 
     it('should create the same cache key when query params are out of order', function (done) {
@@ -53,6 +63,13 @@ describe('Utils', function () {
       var key1 = Utils.createCacheKey(params, reference_query)
       var key2 = Utils.createCacheKey(params4, reference_query)
       key1.should.equal(key2)
+      done()
+    })
+
+    it('should create a difference cache key when there is different where clauses', function (done) {
+      var key1 = Utils.createCacheKey(params, test_query2)
+      var key2 = Utils.createCacheKey(params, test_query3)
+      key1.should.not.equal(key2)
       done()
     })